Removed Offset H1, Refactored Posts Feed Into Partial
Removed the h1 I was using to test the h1, removed the use of modulo function. I have no idea what I was doing the last time I used it, but it worked on that site 🤷‍♂️ The post feed now accepts a variable, so it is shared by home & blog now. Fixed minor offset issue with sticky header